Change in ownership (#1)
* Change references from `cedar-team` to `mmabey`

* Version bump - change in ownership

* Specify Poetry version in workflow

Poetry version 1.6.0 dropped support for Python 3.7, which is causing
the automated tests to fail. Work is already underway to remove
pyredox's support for Python 3.7 as part of a larger effort to upgrade
to Pydantic v2, but that's not ready yet. For now, as part of the change
in ownership, I'm just specifying the Poetry version that still supports
3.7 to get tests to pass.